requirements.txt
anyio==4.9.0
APScheduler==3.11.0
asyncpg==0.30.0
certifi==2025.1.31
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
psycopg2-binary==2.9.10
python-telegram-bot==22.0
sniffio==1.3.1
sqlparse==0.5.3
tabulate==0.9.0
typing_extensions==4.13.1
tzlocal==5.3.1
yoyo-migrations==9.0.0
zipp==3.21.0
